Add export macro for private classes with unittest.
Summary:
Otherwise I get
    undefined reference to `vtable for MailCommon::KMFilterAccountList'
when linking filteractionmissingaccountdialogtest.
Must be a consequence of the recent changes in cmake git (autogen feature).

There are a ton more files that are included directly rather than the
classes being exported though (which can make refactorings harder, e.g.
when new cpp files are added as dependencies of the existing ones,
suddenly it becomes complex to know what to compile in).

Should we go fully in the direction of exporting the stuff we need
when building tests, and compiling these cpp files only once, as a result?

Test Plan: Now it builds with clang

Reviewers: mlaurent, ervin, dvratil

Reviewed By: dvratil

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D4708