FolderSettings: Fix "no resource found in collection" assert.
Summary:
Use collectionFromId on demand when the resource is needed and wasn't
set in the collection passed to FolderSettings::forCollection.
For instance this happens in KMail::Util::mailingListsFromMessage
which uses item.parentCollection(), where the resource isn't set.

We do this on demand rather than in setCollection() because
in many cases (when other methods of FolderSettings are being called)
we don't actually need the resource.

Test Plan:
Reply followed by Forward-as-attachment would often hit
the assert for me, especially when using the toolbar buttons (probably
unrelated?).

Reviewers: mlaurent, dvratil, vkrause

Reviewed By: vkrause

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D14010