Add special rules for collection order for Unified Mailboxes agent
Summary:
Regardless of user's sort order preferences, always show the unified mailboxes
agent first.

The reason is that this is not a real mail resource, so it's hidden from
various lists like the account sorting list etc. and I believe it is
safe to assume that anyone who wants to use unified folders will want
to have them at the top (it's all about easy access to folders after
all).

Reviewers: mlaurent

Reviewed By: mlaurent

Subscribers: ngraham, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D15064