Folder Select dialogue: Remember the expansion state between uses
If the user has many folders, then always showing the fully expanded
tree when this dialogue opens may need a lot of scrolling to find the
required folder.  Remembering the expansion state allows irrelevant
parts of the tree to be hidden.  The preselected folder (if there is
one) or the filter results will always be shown expanded.