Load filters with interactive = false except when importing
Prevents the "filter account missing" dialog - which
unconditionally and permanently deletes the filter if the right
agent just doesn't exist (yet, can be startup order problem) -
when KMail loads its filters during startup.
If something goes wrong, it's a much better failure mode to
silently fail to apply a given filter for that session than to
delete the filter outright.

The commit message of the interactivity feature (from 2011) said:
"Allow to select account when we import filter": it was meant for
filter import as an explicit user action - and in that case, it's
okay to delete the filter if nothing to apply it on was selected.
The use of FilterImporterExporter in FilterManager inadvertently
brought interactive = true with it, destructive side effects and
all.