Load filters with interactive = false except when importing
Prevents the "filter account missing" dialog - which
unconditionally and permanently deletes the filter if the right
agent just doesn't exist (yet, can be startup order problem) -
when loading the filters using FilterManager in KMail.
If something goes wrong, it's a much better failure mode to
silently fail to apply a given filter for that session than to
delete the filter outright.

The commit message of the interactivity feature (from 2011) said:
"Allow to select account when we import filter", and the use of
FilterImporterExporter in FilterManager brought interactive = true
with it as an unfortunate side effect, apparently.