CollectionExpiryWidget: Ensure that the "invalid date" setting is saved
It was not being copied, either when converting from an
ExpireCollectionAttribute in convertFromExpireCollectionAttribute(),
or converting to the attribute in CollectionExpiryWidget::save().

A helper function to perform the first conversion was introduced
in commit 8b7b30ec, but the problem existed before that.

Note that there is a confusing inversion in variable naming here: the
CollectionExpirySettings has member 'expiryMessagesWithInvalidDate'
but an ExpireCollectionAttribute has getter and setter functions for
expireMessagesWithValidDate.  Converting here inverts the setting so
as to agree with those names.

Similarly, in ExpireJob the member variable mExpireMessagesWithoutInvalidDate
is set from the attribute expireMessagesWithValidDate().  This
effectively means the same thing, but the double negative is confusing.